+++ packages/Helm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Helm changelog
 
+## 2.5.3 (WIP)
+
+### Bug fixes
+
+* Fix test Helm HelmHelper.removeGaps
+* Fix using IMonomerLibHelper.loadMonomerLibForTests
+
 ## 2.5.2 (2024-09-25)
 
 Bump dependencies, bio lib